requirements.txt
Django>=4.2,<4.3
psycopg2>=2.8.6,<2.9
selenium==4.11.2
requests==2.31.0
django-tinymce==3.6.1
django-crispy-forms==2.0
pillow==10.0.0
coverage==7.3.1
